Updating files with proper names, renamed singleforsubmit to submitformsingleinput
