+++ src/client/components/SubmitFormSingleInputComponent/SubmitFormSingleInput.style.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Rammetto+One&family=Roboto&display=swap');
-
 .m-top {
   margin-top: 1%;
 }
